agent-runtime/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "neuralis-agent-runtime"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
inference = ["llama-cpp-python>=0.2.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
